ENH: moved stuff over to motor.py and reworked methods
